fix: data access layer audit — C1–C6, H1, H4, H5, M3
Fixes 10 issues from the data repository audit (#5680):

C1: Whitelist-based ORDER BY guard in all 10 query builders (SQL injection)
C2: Regex validation for sequence names in EvidenceRepository (SQL injection)
C3: @Transactional + idempotent UPDATE guards in HearingRepository/DBRepository
C4: CompletableFuture three-wave parallelization in CaseRepository.enrichCaseCriteria
C5: Parenthesise date-range OR clauses in CaseQueryBuilder (operator precedence)
C6: COUNT(*) OVER() window function replaces double-query count in CaseRepository
H1: Move advocate/office validation out of ApplicationRepository into ApplicationValidator
H4: Inline SQL moved to HearingQueryBuilder.getHearingsWithMultipleFilingNumbersQuery + LIMIT 500 + JSONB index migration
H5: Replace SELECT * with explicit column list in HearingQueryBuilder (HEARING_SELECT_COLUMNS)
M3: Standardise Pagination.limit/offSet from Double to Integer across 12 models, 9 query builders, all callers and tests

+++ backend/dristi-services/advocate-office-management/src/main/java/digit/web/models/Pagination.java
@@ -24,10 +24,10 @@ public class Pagination {
 
     @JsonProperty("limit")
     @DecimalMax("100")
-    private Double limit = 10d;
+    private Integer limit = 10;
 
     @JsonProperty("offSet")
-    private Double offSet = 0d;
+    private Integer offSet = 0;
 
     @JsonProperty("totalCount")
     private Double totalCount = null;
