+++ backend/dristi-services/analytics/src/main/java/org/pucar/dristi/web/models/taskManagement/Pagination.java
@@ -22,10 +22,10 @@ import org.springframework.validation.annotation.Validated;
 public class Pagination {
     @JsonProperty("limit")
     @DecimalMax("100")
-    private Double limit = 10d;
+    private Integer limit = 10;
 
     @JsonProperty("offSet")
-    private Double offSet = 0d;
+    private Integer offSet = 0;
 
     @JsonProperty("totalCount")
     private Double totalCount = null;
