+++ backend/dristi-services/digitalized-documents/src/main/java/digit/repository/querybuilder/DigitalizedDocumentQueryBuilder.java
@@ -34,10 +34,10 @@ public class DigitalizedDocumentQueryBuilder {
 
     public String addPaginationQuery(String query, Pagination pagination, List<Object> preparedStatementList, List<Integer> preparedStatementArgList) {
         preparedStatementList.add(pagination.getLimit());
-        preparedStatementArgList.add(Types.DOUBLE);
+        preparedStatementArgList.add(Types.INTEGER);
 
         preparedStatementList.add(pagination.getOffSet());
-        preparedStatementArgList.add(Types.DOUBLE);
+        preparedStatementArgList.add(Types.INTEGER);
 
         return query + " LIMIT ? OFFSET ?";
     }
